+++ shared/include/models/object.hpp
@@ -9,7 +9,7 @@ class Object {
     using Metadata =
         std::unordered_map<std::string, std::variant<int, double, std::string>>;
     Object() = default;
-    Object(const std::string& name, double x, double y, const std::string& type, double creation_time)
+    Object(const std::string& name, double x, double y, const std::string& type, unsigned long creation_time)
         : name_(name), x_(x), y_(y), type_(type), creation_time_(creation_time) {}
  public:
     // Getters
@@ -17,7 +17,7 @@ class Object {
     [[nodiscard]] double GetX() const { return x_; }
     [[nodiscard]] double GetY() const { return y_; }
     [[nodiscard]] std::string GetType() const { return type_; }
-    [[nodiscard]] double GetCreationTime() const { return creation_time_; }
+    [[nodiscard]] unsigned long GetCreationTime() const { return creation_time_; }
     [[nodiscard]] Metadata& GetMetadata() { return metadata_; }
 
     // Setters
@@ -25,7 +25,7 @@ class Object {
     void SetX(double x) { x_ = x; }
     void SetY(double y) { y_ = y; }
     void SetType(const std::string& type) { type_ = type; }
-    void SetCreationTime(double creation_time) { creation_time_ = creation_time; }
+    void SetCreationTime(unsigned long creation_time) { creation_time_ = creation_time; }
     void SetMetadata(const Metadata& metadata) { metadata_ = metadata; }
 
  private:
@@ -33,6 +33,6 @@ class Object {
     double x_;
     double y_;
     std::string type_;
-    double creation_time_;
+    unsigned long creation_time_;
     Metadata metadata_;
 };
